+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/DsimTargetDomainObject.kt
@@ -12,8 +12,8 @@ import java.nio.file.Path
 interface DsimTargetDomainObject : TargetDomainObject {
 
     var compileTops: List<String>
-    var extraIncludeDirs: List<Path>
     var extraFiles: List<Path>
+    var extraIncludeDirs: List<Path>
     var dpiLibs: List<Path>
 
     fun sim(configure: DsimSimDomainObject.() -> Unit)