+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/IverilogTargetDomainObject.kt
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.domain
+
+/**
+ * Interface for domain objects that configure an Icarus Verilog target.
+ */
+interface IverilogTargetDomainObject : TargetDomainObject {
+
+    var top: String
+}