+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/VerikDomainObjectImpl.kt
@@ -35,6 +35,13 @@ class VerikDomainObjectImpl : VerikDomainObject {
         targetDomainObjects.add(domainObject)
     }
 
+    override fun iverilog(configure: IverilogTargetDomainObject.() -> Unit) {
+        val action = Action<IverilogTargetDomainObject> { it.configure() }
+        val domainObject = IverilogTargetDomainObjectImpl()
+        action.execute(domainObject)
+        targetDomainObjects.add(domainObject)
+    }
+
     override fun vivado(configure: VivadoTargetDomainObject.() -> Unit) {
         val action = Action<VivadoTargetDomainObject> { it.configure() }
         val domainObject = VivadoTargetDomainObjectImpl()